flask/requirements.txt
Flask
Redis
flask-sqlalchemy
flask-login
flask-migrate
flask-bootstrap
Flask-WTF
mysqlclient
